requirements.txt
fastapi==0.112.0
uvicorn==0.30.5
pydantic==2.7.2
httpx==0.27.0
#psycopg2==2.9.9
psycopg2-binary==2.9.9
pytest==8.2.2
flake8==7.1.1
python-dotenv==1.0.1